+++ examples/cpp/otel/ostream/BUILD
@@ -14,34 +14,16 @@
 
 licenses(["notice"])
 
-package(
-    default_visibility = ["//examples/cpp/otel:__subpackages__"],
-)
-
-cc_library(
-    name = "util",
-    srcs = ["util.cc"],
-    hdrs = ["util.h"],
-    defines = ["BAZEL_BUILD"],
-    deps = [
-        "//:grpc++",
-        "//:grpc++_reflection",
-        "//examples/protos:helloworld_cc_grpc",
-        "@io_opentelemetry_cpp//sdk/src/metrics",
-    ],
-)
-
 cc_binary(
     name = "greeter_callback_client",
     srcs = ["greeter_callback_client.cc"],
     defines = ["BAZEL_BUILD"],
     deps = [
-        "util",
-        "//:grpc++",
         "//:grpcpp_otel_plugin",
+        "//examples/cpp/otel:util",
         "@com_google_absl//absl/flags:flag",
         "@com_google_absl//absl/flags:parse",
-        "@io_opentelemetry_cpp//exporters/prometheus:prometheus_exporter",
+        "@io_opentelemetry_cpp//exporters/ostream:ostream_metric_exporter",
         "@io_opentelemetry_cpp//sdk/src/metrics",
     ],
 )
@@ -51,13 +33,12 @@ cc_binary(
     srcs = ["greeter_callback_server.cc"],
     defines = ["BAZEL_BUILD"],
     deps = [
-        "util",
-        "//:grpc++",
         "//:grpcpp_otel_plugin",
+        "//examples/cpp/otel:util",
         "@com_google_absl//absl/flags:flag",
         "@com_google_absl//absl/flags:parse",
         "@com_google_absl//absl/strings:str_format",
-        "@io_opentelemetry_cpp//exporters/prometheus:prometheus_exporter",
+        "@io_opentelemetry_cpp//exporters/ostream:ostream_metric_exporter",
         "@io_opentelemetry_cpp//sdk/src/metrics",
     ],
 )
